Implement refresh mechanism
The auth data was extended by the expire time, this was ignored in the
past. Other api functions were altered so that authentication is the last
argument. With that change they can easily be used in the new function
which enhances a api function with auto refresh of the authentication.